refactor(commands): reorganize OpenAPI and Services generation commands into a new namespace for better structure; remove old command files
